+++ src/components/common/RegisterInput/index.tsx
@@ -1,5 +1,7 @@
 import type { ComponentProps } from "react";
 import { forwardRef } from "react";
+import type { Path, UseFormRegister } from "react-hook-form";
+import type { CompanyInfoStateType } from "@/schemas/companyInfo.ts";
 import styled from "@emotion/styled";
 import { palette } from "@/styles/palette";
 import { typo } from "@/styles/typo";
@@ -9,20 +11,39 @@ interface InputProps extends ComponentProps<"input"> {
   height?: number;
   placeholder: string;
   type?: string;
+  label?: Path<CompanyInfoStateType>;
+  register?: UseFormRegister<CompanyInfoStateType>;
 }
 
-const RegisterInput = forwardRef<HTMLInputElement, InputProps>(function Input(
-  { width, height = 46, type = "text", placeholder, ...props }: InputProps,
-  inputRef,
-) {
+/**
+ * @param width : input width
+ * @param height : input height
+ * @param placeholder : input placeholder
+ * @param type : input type
+ * @param ref : input ref
+ * @param label : input react-hook-form label
+ * @param register : input react-hook-form register
+ * @param props : input other props
+ * @description : 정보 등록 및 수정 페이지에서 활용하는 Input 컴포넌트입니다. form 구성 시 register, label를 props로 전달해주세요.
+ * @returns
+ */
+const RegisterInput = forwardRef<HTMLInputElement, InputProps>(function Input({
+  width,
+  height = 46,
+  placeholder,
+  type = "text",
+  label,
+  register,
+  ...props
+}: InputProps) {
   return (
     <StyleInputWrapper>
       <StyleInput
         widthProps={width}
         heightProps={height}
-        ref={inputRef}
         placeholder={placeholder}
         type={type}
+        {...(register && label ? register(label) : {})}
         {...props}
       />
     </StyleInputWrapper>
