+++ src/pages/chatting/Chatting.tsx
@@ -6,6 +6,7 @@ import styled from "@emotion/styled";
 import * as Stomp from "@stomp/stompjs";
 import { axiosAPI } from "@/apis/axios";
 import type { Messages } from "@/apis/chatting/chattingType";
+import MessageArea from "@/pages/chatting/components/messageArea";
 import ProfileSheet from "@/components/common/BottomSheet/ProfileSheet";
 import { FlexBox } from "@/components/common/Flexbox";
 import GradationBackground from "@/components/common/GradationBackground";
@@ -13,7 +14,6 @@ import PageContainer from "@/components/common/PageContainer";
 import PageHeader from "@/components/common/PageHeader";
 import Spacing from "@/components/common/Spacing";
 import TextArea from "@/components/common/TextArea";
-import MessageArea from "@/components/messageArea";
 import { useModal } from "@/hooks/useModal";
 import useToast from "@/hooks/useToast";
 import { palette } from "@/styles/palette";
