+++ src/pages/home/components/AvatarGroup.tsx
