+++ src/pages/home/components/Tip.tsx
